Fix the code to match checkstyle
