+++ build.gradle
@@ -35,7 +35,7 @@ test {
 }
 
 application {
-    mainClass.set("Codsworth.Codsworth")
+    mainClass.set("codsworth.Codsworth")
 }
 
 shadowJar {
