+++ src/main/java/codsworth/CodsworthExceptions/CodsworthOutOfBoundsException.java
@@ -1,4 +1,4 @@
-package Codsworth.CodsworthExceptions;
+package codsworth.CodsworthExceptions;
 
 public class CodsworthOutOfBoundsException extends RuntimeException {
     @Override
