+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ ________________________________________________________
 | | |___| (_) | (_| \__ \\ V  V / (_) | |  | |_| | | | |
 |  \_____\___/ \__,_|___/ \_/\_/ \___/|_|   \__|_| |_| |
 ________________________________________________________
-Hello there, my name is Codsworth.Codsworth
+Hello there, my name is Codsworth
 What can I do for you?
 ________________________________________________________
 ________________________________________________________
@@ -48,6 +48,6 @@ ________________________________________________________
 | | |___| (_) | (_| \__ \\ V  V / (_) | |  | |_| | | | |
 |  \_____\___/ \__,_|___/ \_/\_/ \___/|_|   \__|_| |_| |
 ________________________________________________________
-Thank you for using Codsworth.Codsworth
+Thank you for using Codsworth
 Hope to see you soon!
 ________________________________________________________
