+++ text-ui-test/runtest.sh
@@ -20,7 +20,7 @@ then
 fi
 
 # run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ../bin Codsworth.Codsworth < input.txt > ACTUAL.TXT
+java -classpath ../bin codsworth.Codsworth < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
 cp EXPECTED.TXT EXPECTED-UNIX.TXT
